requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==4.2.17
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
Markdown==3.8
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
